GENERIC_READ was missing from the CreateFile call when formatting the device; fixed compile warning of implicit casting
git-svn-id: https://svn.code.sf.net/p/szte-wsn/code@2954 a4e16f3a-606c-4714-833f-ed604d7aca08

+++ private/ali/RawToCType/Utility.cpp
@@ -248,7 +248,7 @@ int32_t cast_to_int32(int64_t size) {
 
 	throw_if_larger_than_2GB(size);
 
-	return size;
+	return static_cast<int32_t>(size);
 }
 
 void replace(string& s, const char old, const char new_char) {
